Export iteration functions and move them to .cpp file
