+++ PolyEngine/Core/Src/pe/core/utils/IterationUtils.hpp
@@ -4,19 +4,23 @@
 
 namespace pe::core::utils
 {
-    size_t prevIdxCircularBuffer(size_t head, size_t size, size_t capacity, size_t idx)
-    {
-        const size_t sizeWithGuard = size + 1;
-        const size_t listIdx = (idx + capacity - head) % capacity;
-        const size_t sizeCorrectedPrevOffset = (listIdx + sizeWithGuard - 1) % sizeWithGuard;
-        return (head + sizeCorrectedPrevOffset) % capacity;
-    }
+    /** Obtains previous physical index in a circular buffer.
+     * 
+     * @param head Buffer head
+     * @param size Buffer actual size
+     * @param capacity Buffer total capacity
+     * @param idx Buffer current index
+     * @return Previous index in circular buffer.
+     */
+    size_t CORE_DLLEXPORT prevIdxCircularBuffer(size_t head, size_t size, size_t capacity, size_t idx);
     
-    size_t nextIdxCircularBuffer(size_t head, size_t size, size_t capacity, size_t idx)
-    {
-        const size_t sizeWithGuard = size + 1;
-        const size_t listIdx = (idx + capacity - head) % capacity;
-        const size_t sizeCorrectedNextOffset = (listIdx + sizeWithGuard + 1 ) % sizeWithGuard;
-        return (head + sizeCorrectedNextOffset) % capacity;
-    }
+    /** Obtains next physical index in a circular buffer.
+     * 
+     * @param head Buffer head
+     * @param size Buffer actual size
+     * @param capacity Buffer total capacity
+     * @param idx Buffer current index
+     * @return Next index in circular buffer.
+     */
+    size_t CORE_DLLEXPORT nextIdxCircularBuffer(size_t head, size_t size, size_t capacity, size_t idx);
 }
